backend/requirements.txt
openai
pinecone
requests
beautifulsoup4
python-dotenv
google-search-results
fastapi
uvicorn